# Laravel-API-Server-master/Laravel-API-Server-master/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Api Juegos",
        "description": "Documentación de todos los endpoints de un aplicación que consiste en un CRUD de Juegos que contiene relaciones, donde se trabaja con el envió de archivos y también contiene filtros de búsqueda sobre resultados de los juegos que existen insertados, se puede registrar usuarios y también contiene autenticación por SANCTUM para poder realizar la acción de algunos endpoints.<br><br><a href='https://github.com/JAVI-CC/Laravel-API-Server' target='_blank'>https://github.com/JAVI-CC/Laravel-API-Server</a><br><br><a href='https://github.com/JAVI-CC/Laravel-API-Client' target='_blank'>https://github.com/JAVI-CC/Laravel-API-Client</a><br><br><a href='https://github.com/JAVI-CC/VUE3-API-client' target='_blank'>https://github.com/JAVI-CC/VUE3-API-client</a>",
        "version": "1.0.1"
    },
    "paths": {
...
        "/api/juegos": {
            "get": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Obtener todos los juegos",
                "description": "Muestra todos los registros de juegos en formato JSON",
                "operationId": "getAllJuegos",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Insertar un juego",
                "description": "Insertar el registro de un juego nuevo con parametros.",
                "operationId": "addJuego",
                "parameters": [
                    {
                        "name": "Juego",
                        "in": "query",
                        "description": "{nombre, descripcion, desarrolladora, fecha, generos, imagen}",
                        "required": true,
                        "schema": {
                            "properties": {
                                "nombre": {
                                    "$ref": "#/components/schemas/Juego/properties/nombre"
                                },
                                "descripcion": {
                                    "$ref": "#/components/schemas/Juego/properties/descripcion"
                                },
                                "desarrolladora": {
                                    "$ref": "#/components/schemas/Juego/properties/desarrolladora"
                                },
                                "imagen": {
                                    "$ref": "#/components/schemas/Juego/properties/imagen"
                                },
                                "fecha": {
                                    "$ref": "#/components/schemas/Juego/properties/fecha"
                                },
                                "generos[0]": {
                                    "type": "string",
                                    "example": "accion"
                                },
                                "generos[1]": {
                                    "type": "string",
                                    "example": "first-person-shooter"
                                },
                                "generos[2]": {
                                    "type": "string",
                                    "example": "multijugador"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Se ha creado correctamente"
                    },
                    "220": {
                        "description": "No se cumple todos los requisitos"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "SANCTUM": []
                    }
                ]
            }
        },
        "/api/juegos/paginate": {
            "get": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Obtener los juegos por una paginación de X elementos",
                "description": "Muestra los juegos a través de una paginación",
                "operationId": "getPaginateJuegos",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numero de la paginación",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": "2"
                        }
                    },
                    {
                        "name": "items",
                        "in": "query",
                        "description": "Numero de juegos por página",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "8"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "ordenar paginación de juegos (no es obligatorio). Opciones: [nombreAsc, nombrDesc, fechaAsc, fechaDesc]",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/juegos/random": {
            "get": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Obtener una lista de X juegos aleatorios",
                "description": "Muestra el resultado de X juegos de forma aleatoria",
                "operationId": "getRandomJuegos",
                "parameters": [
                    {
                        "name": "items",
                        "in": "query",
                        "description": "Numero de juegos",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "6"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/juegos/{slug}": {
            "get": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Obtener un juego",
                "description": "Muestra la información de un juego especifico segun el valor del parametro slug.",
                "operationId": "getJuego",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Url del nombre del juego",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "test123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/juegos/edit": {
            "post": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Actualizar un juego",
                "description": "Actulizar un juego ya existente",
                "operationId": "editJuego",
                "parameters": [
                    {
                        "name": "Juego",
                        "in": "query",
                        "description": "{nombre, descripcion, desarrolladora, fecha, generos, slug, imagen}",
                        "required": true,
                        "schema": {
                            "properties": {
                                "nombre": {
                                    "$ref": "#/components/schemas/Juego/properties/nombre"
                                },
                                "descripcion": {
                                    "$ref": "#/components/schemas/Juego/properties/descripcion"
                                },
                                "desarrolladora": {
                                    "$ref": "#/components/schemas/Juego/properties/desarrolladora"
                                },
                                "imagen": {
                                    "$ref": "#/components/schemas/Juego/properties/imagen"
                                },
                                "fecha": {
                                    "$ref": "#/components/schemas/Juego/properties/fecha"
                                },
                                "generos[0]": {
                                    "type": "string",
                                    "example": "aventura"
                                },
                                "generos[1]": {
                                    "type": "string",
                                    "example": "rpg-de-accion"
                                },
                                "generos[2]": {
                                    "type": "string",
                                    "example": "multijugador"
                                },
                                "slug": {
                                    "$ref": "#/components/schemas/Juego/properties/slug"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "220": {
                        "description": "No se cumple todos los requisitos"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "SANCTUM": []
                    }
                ]
            }
        },
        "/api/juegos/delete/{slug}": {
            "delete": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Eliminar un juego",
                "description": "Elimina un juego especifico segun el valor del parametro slug.",
                "operationId": "deleteJuego",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Url del nombre del juego",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "test123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "SANCTUM": []
                    }
                ]
            }
        },
        "/api/juegos/filter/search": {
            "post": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Busqueda",
                "description": "Busqueda por nombres de juegos y también ordena el resultado de distintas maneras.",
                "operationId": "filterJuego",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Nombre del juego que quieres buscar",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "dark"
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Ordenar por diferente tipo. Ejemplos ['nombre', 'descripcion', 'fecha'] ",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "fecha"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "ordenar de forma 'ASC' o 'DESC'",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "DESC"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "220": {
                        "description": "No se cumple todos los requisitos"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "SANCTUM": {
                "type": "apiKey",
                "description": "Clave API de tipo Bearer Token para poder utilizar todas las peticiones.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}